refactor: reorganize Account classes into utils/accounts/ subfolder
